Changed waiter to calculate primes due to performance issues with repeatedly reading system time

+++ Stage_demo/Stage_demo.cpp
@@ -36,7 +36,7 @@ int _tmain(int argc, _TCHAR* argv[])
 	int SPHERES = 5;
 	//Säikeiden määrä
 	uint32_t THREADS = 16;
-	double WAIT = 0.0;
+	int WAIT = 0;
 	std::string configfile;
 	std::ifstream configStream("config.ini", std::ios::in);
 	//Luetaan parametrit konfiguraatiotiedostosta
@@ -90,8 +90,8 @@ int _tmain(int argc, _TCHAR* argv[])
 			//WAIT-parametri
 			else if (start == "WAIT"){
 				try{
-					WAIT = std::stod(end);
-					if (WAIT < 0.0) WAIT = 0.0;
+					WAIT = std::stoi(end);
+					if (WAIT < 0) WAIT = 0;
 				}
 				catch (...){
 					std::cerr << "Error parsing configuration parameter WAIT" << std::endl;
